fix: don't save state of SAFMode="single" files
Files opened with `SAFMode="single"` are accessed via a SAF intent, which provides only temporary access to their URIs. Saving the state for these files can cause issues on the next app startup, as the temporary access is revoked and the app will be unable to access the file again. This change prevents saving the state for such files, avoiding potential errors and other related issues.